% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\docType{data}
\name{cases}
\alias{cases}
\title{Cases Dataset}
\format{
\subsection{\code{cases}}{

A tibble with 2970 rows and 14 columns:
\describe{
\item{Id}{Person ID}
\item{Event}{Binary indicator for the outcome}
\item{ex}{binary indicator for the exposure}
\item{sex}{person sex}
\item{age}{person age}
\item{day}{day of follow-up, the last day is the case period}
}
}
}
\source{
\url{https://github.com/lan-k/Casecrossover}
}
\usage{
cases
}
\description{
A dataset containing  cases for a case-crossover study from simulated data
}
\keyword{datasets}
