% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_tvc.R
\docType{data}
\name{cases_tvc}
\alias{cases_tvc}
\title{Cases Dataset with binary time-varying confounder}
\format{
\subsection{\code{cases_tvc}}{

A tibble with 2970 rows and 14 columns:
\describe{
\item{Pt_ID}{Person ID}
\item{event}{Binary indicator for the outcome}
\item{ex}{binary indicator for the exposure}
\item{z}{binary time-varying confounder}
\item{period}{period number}
}
}
}
\source{
\url{https://github.com/lan-k/Casecrossover}
}
\usage{
cases_tvc
}
\description{
A dataset containing  cases for a case-crossover study from simulated data
}
\keyword{datasets}
