% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casetimecontrols.R
\docType{data}
\name{casetimecontrols}
\alias{casetimecontrols}
\title{Casetimecontrols Dataset}
\format{
\subsection{\code{casetimecontrols}}{

A tibble with 56430 rows and 14 columns:
\describe{
\item{Id}{Person ID}
\item{Event}{Binary indicator for the outcome}
\item{ex}{binary indicator for the exposure}
\item{sex}{person sex}
\item{age}{person age}
\item{day}{day of follow-up, the last day is the case period}
}
}
}
\source{
\url{https://github.com/lan-k/Casecrossover}
}
\usage{
casetimecontrols
}
\description{
A dataset containing  cases and time controls for a case-time-control study from simulated data
}
\keyword{datasets}
