% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXO_tc_funcs.R
\name{CXO_tc_wt}
\alias{CXO_tc_wt}
\title{Weighted conditional logistic regression for case-time-control study}
\usage{
CXO_tc_wt(data, exposure, event, Id, tvc = NULL)
}
\arguments{
\item{data}{input dataframe}

\item{exposure}{exposure variable}

\item{event}{outcome variable}

\item{Id}{person ID}

\item{tvc}{time-varying confounder (optional)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    weighted regression object from clogit
}\if{html}{\out{</div>}}
}
\description{
Weighted conditional logistic regression for case-time-control study
}
\examples{
data(casetimecontrols)
ctcfit <- CXO_tc_wt(data=casetimecontrols,exposure=ex,event=Event,Id=Id)
summary(ctcfit)
}
