% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXO_funcs.R
\name{CXO_wt}
\alias{CXO_wt}
\title{Weighted Conditional Logistic Regression for CXO study}
\usage{
CXO_wt(data, exposure, event, Id)
}
\arguments{
\item{data}{input dataframe}

\item{exposure}{name of exposure variable}

\item{event}{name of outcome variable}

\item{Id}{person ID}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      weighted regression object from clogit
}\if{html}{\out{</div>}}
}
\description{
Weighted Conditional Logistic Regression for CXO study
}
\examples{
data(cases)
cfit <- CXO_wt(data=cases, exposure=ex, event=Event, Id=Id)
summary(cfit)
}
