% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXO_tc_funcs.R
\name{.CI_tc_boot}
\alias{.CI_tc_boot}
\title{Internal function for Bootstrapped CIs}
\usage{
.CI_tc_boot(data, ii, exposure, event, Id)
}
\arguments{
\item{data}{input dataframe}

\item{ii}{index for bootstrapping}

\item{exposure}{exposure variable}

\item{event}{outcome}

\item{Id}{person ID}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      coefficient of bootstrapped clogit
}\if{html}{\out{</div>}}
}
\description{
Internal function for Bootstrapped CIs
}
