% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXO_funcs.R
\name{SCL_bias}
\alias{SCL_bias}
\title{Estimate bias due to conditional logistic regression in CXO study}
\usage{
SCL_bias(data, exposure, event, Id)
}
\arguments{
\item{data}{input dataframe}

\item{exposure}{exposure variable}

\item{event}{outcome variable}

\item{Id}{person Id for strata}
}
\value{
A list containing:\tabular{ll}{
\code{scl} \tab Standard Conditional logistic regression OR \cr
\tab \cr
\code{mh_or} \tab Mantel_Haenszel OR \cr
\tab \cr
\code{bias} \tab \% difference between the CL and MH OR \cr
}
}
\description{
Estimate bias due to conditional logistic regression in CXO study
}
\examples{
data(cases)
SCL_bias(data=cases, exposure=ex, event=Event, Id=Id)
}
