% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXO_funcs.R
\name{CXO_wt_boot}
\alias{CXO_wt_boot}
\title{Weighted conditional logistic regression for case-crossover with bootstrapped CIs}
\usage{
CXO_wt_boot(data, exposure, event, Id, B = 500, normal = TRUE)
}
\arguments{
\item{data}{input dataframe}

\item{exposure}{exposure variable}

\item{event}{outcome variable}

\item{Id}{person Id}

\item{B}{number of bootstrapped replicates, default is 500. Minimum 200 recommended}

\item{normal}{use normal approximation for bootstrapped CI, default is TRUE
normal = F uses 2.5\% and 97.\% quantiles for 95\% CI}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    OR estimate and bootstrapped 95\% CI
}\if{html}{\out{</div>}}
}
\description{
Weighted conditional logistic regression for case-crossover with bootstrapped CIs
}
\examples{
# Chop ----------------------------------------------------------------------
data(cases)
try(CXO_wt_boot(data=cases, exposure=ex, event=Event, Id = Id, B=3))
# It is recommended that at least 500 bootstrap replications are used
# Unchop ----------------------------------------------------------------------
\dontrun{
CXO_wt_boot(data=cases, exposure=ex, event=Event, Id = Id, B=1000)
}
}
