% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXOwt-package.R
\docType{package}
\name{CXOwt-package}
\alias{CXOwt}
\alias{CXOwt-package}
\title{CXOwt: Weighted Conditional Logistic Regression for Case-Crossover and Case-Time Control Studies}
\description{
Produce weighted OR estimates with reduced bias in case-crossover and case-time control studies.
}
\author{
\strong{Maintainer}: Lan Kelly \email{Lan.Kelly@unisa.edu.au} (\href{https://orcid.org/0000-0002-7691-9289}{ORCID})

}
\keyword{internal}
