% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXO_funcs.R
\name{SCL_bias}
\alias{SCL_bias}
\title{Estimate bias due to conditional logistic regression in CXO study}
\usage{
SCL_bias(data, exposure, event, Id)
}
\arguments{
\item{data}{input dataframe}

\item{exposure}{exposure variable}

\item{event}{outcome variable}

\item{Id}{person Id for strata}
}
\value{
list with bias, CL and MH estimates
bias \% difference between the CL and MH OR
scl Standard Conditional logistic regression OR
mh_or Mantel_Haenszel OR
}
\description{
Estimate bias due to conditional logistic regression in CXO study
}
\examples{
data(cases)
SCL_bias(data=cases, exposure=ex, event=Event, Id=Id)
}
